Fix sign-conversion warnings in headers

+++ include/fmt/chrono.h
@@ -250,7 +250,7 @@ struct chrono_formatter {
   void write(int value, int width) {
     typedef typename int_traits<int>::main_type main_type;
     main_type n = to_unsigned(value);
-    int num_digits = static_cast<int>(internal::count_digits(n));
+    int num_digits = internal::count_digits(n);
     if (width > num_digits)
       out = std::fill_n(out, width - num_digits, '0');
     out = format_decimal<char_type>(out, n, num_digits);
